Create a file for test util's functional test file
